requirements.txt
selenium
bibtexparser
python-dotenv
